Move scenario contexts
