+++ src/page-multi-facility/MultiFacilityPage.tsx
@@ -2,9 +2,9 @@ import React from "react";
 import styled from "styled-components";
 
 import SiteHeader from "../components/site-header/SiteHeader";
+import useScenario from "../contexts/scenario-context/useScenario";
 import Loading from "../design-system/Loading";
 import { useLocaleDataState } from "../locale-data-context";
-import useScenario from "../scenario-context/useScenario";
 import CreateBaselineScenarioPage from "./CreateBaselineScenarioPage";
 import MultiFacilityImpactDashboard from "./MultiFacilityImpactDashboard";
