+++ src/page-multi-facility/ScenarioLibraryModal.tsx
@@ -1,6 +1,7 @@
 import React, { useEffect, useState } from "react";
 import styled from "styled-components";
 
+import useScenario from "../contexts/scenario-context/useScenario";
 import { deleteScenario, duplicateScenario, getScenarios } from "../database";
 import Colors from "../design-system/Colors";
 import { DateMMMMdyyyy } from "../design-system/DateFormats";
@@ -11,7 +12,6 @@ import Loading from "../design-system/Loading";
 import Modal, { Props as ModalProps } from "../design-system/Modal";
 import ModalDialog from "../design-system/ModalDialog";
 import PopUpMenu from "../design-system/PopUpMenu";
-import useScenario from "../scenario-context/useScenario";
 import { Scenario } from "./types";
 
 const ModalContents = styled.div`
