+++ src/site-provider/SiteProvider.tsx
@@ -4,6 +4,7 @@ import { ToastProvider } from "react-toast-notifications";
 
 import AppAuth0ClientPromise from "../auth/AppAuth0ClientPromise";
 import { Auth0Provider } from "../auth/react-auth0-spa";
+import { ScenarioProvider } from "../contexts/scenario-context";
 import Toast from "../design-system/Toast";
 import { FeatureFlagsProvider } from "../feature-flags";
 import { LocaleDataProvider } from "../locale-data-context";
@@ -12,7 +13,6 @@ import {
   rtDataReducer,
 } from "../page-multi-facility/FacilityContext";
 import { Facility } from "../page-multi-facility/types";
-import { ScenarioProvider } from "../scenario-context";
 
 // A function that routes the user to the right place after login
 const onRedirectCallback = (appState: any) => {
